requirements.txt
Django==4.2.2
keras==2.12.0
numpy==1.23.5
Pillow==9.5.0
